Modify hll_cnt_rho to handle corner cases.
The function hll_cnt_rho now returns the number of bits of the hash data
type minus one in case of a 0 hash. This is because returning the actual
number of bits (32 in case of 32 bit hashes), when calling hll_cnt_size
we got the wrong result. That is, when computing the denominator of the
expression to compute the sizewe had a shift of 32 positions, resulting
in a 0 denominator value.